+++ app/src/main/java/com/mobileprogramming/finsheet/domain/model/DashboardData.kt
@@ -1,15 +1,15 @@
 package com.mobileprogramming.finsheet.domain.model
 
 data class DashboardData(
-    val totalBalance: Long,
-    val incomeThisMonth: Long,
-    val expenseThisMonth: Long,
+    val totalBalance: Double,
+    val incomeThisMonth: Double,
+    val expenseThisMonth: Double,
     val categoryExpensesToday: List<CategoryExpenseModel>,
     val categoryExpensesThisWeek: List<CategoryExpenseModel>,
     val categoryExpensesThisMonth: List<CategoryExpenseModel>,
-    val totalExpenseToday: Long,
-    val totalExpenseThisWeek: Long,
-    val totalExpenseThisMonth: Long,
+    val totalExpenseToday: Double,
+    val totalExpenseThisWeek: Double,
+    val totalExpenseThisMonth: Double,
     val monthlyBudgets: List<BudgetProgressModel>
 )
 
@@ -18,7 +18,7 @@ data class CategoryExpenseModel(
     val categoryName: String,
     val icon: String? = null,
     val color: String? = null,
-    val totalAmount: Long
+    val totalAmount: Double
 )
 
 data class BudgetProgressModel(
@@ -27,6 +27,6 @@ data class BudgetProgressModel(
     val categoryId: String,
     val icon: String? = null,
     val color: String? = null,
-    val limitAmount: Long,
-    val usedAmount: Long
+    val limitAmount: Double,
+    val usedAmount: Double
 )
